flora_flow_landing/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-axes==7.0.1
django-cors-headers==4.6.0
django-environ==0.11.2
factory_boy==3.3.1
Faker==33.1.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
requests
Telethon==1.37.0
python-telegram-bot==21.7